Rework billing plans: flat per-click pricing with a -15% discount badge
Replace the 4 flat-price plans with 3 tiers priced off a single rate
(CLICK_PRICE = 0.0058 UAH/click in app/core/plans.py): 50k, 500k and
2,000,000 credits. Each card now shows the discounted price plus the
struck-through original and a "-15%" badge (DISCOUNT_PERCENT), computed
via base_price()/final_price() so changing the rate or discount updates
every plan automatically. Dropped the bullet-point feature lists per
request — cards are just name, price and credit count now.

payment_orders.amount was INTEGER; discounted prices land on half-hryvnia
values (e.g. 246.50), so widened it to NUMERIC(10,2) with a guarded
Postgres migration in main.py for databases where the table already
exists from the previous deploy.

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Boolean, DateTime, ForeignKey, JSON, Text, Index
+from sqlalchemy import Column, Integer, Numeric, String, Float, Boolean, DateTime, ForeignKey, JSON, Text, Index
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.database import Base
@@ -89,7 +89,7 @@ class PaymentOrder(Base):
     plan_key = Column(String, nullable=False)
     plan_name = Column(String)
     credits = Column(Integer, nullable=False)
-    amount = Column(Integer, nullable=False)            # у гривнях
+    amount = Column(Numeric(10, 2), nullable=False)     # у гривнях, може бути з копійками
     currency = Column(String, default="UAH")
     status = Column(String, default="pending")          # pending | paid | cancelled
     created_at = Column(DateTime(timezone=True), server_default=func.now())
